webcars/requirements.txt
psycopg2==2.7.4
dj-database-url
Django==2.0.4
django-heroku==0.3.1
gunicorn==19.8.0
psycopg2-binary
django-registration-redux==2.4
django-filter==1.1.0
django-widget-tweaks